+++ app/src/main/java/ru/itis/tinkoff/project/features/profile/data/UserRepositoryImpl.kt
@@ -7,5 +7,5 @@ import ru.itis.tinkoff.project.entity.User
 class UserRepositoryImpl(
     private val api: Api
 ) : UserRepository {
-    override suspend fun getUser(id: Long): User = api.getProfileInfo(id)
+    override suspend fun getUser(id: Int): User = api.getProfileInfo(id)
 }
